Starting to implement continent bonus.
I decided to implement this functionality using the phase "state machine". By adding special cases for add-army when a player owns continents I think I would be able to reuse most of the code.
For now, I changed the game logic code and added a few tests to verify that the implementation is correct. I haven't touched the ui code yet, though.